update delete method and updated method to handle the exception occured.

+++ src/main/java/com/mitchell_international/VehicleApplication/service/VehicleService.java
@@ -11,7 +11,6 @@ import com.mitchell_international.VehicleApplication.repository.VehicleRepositor
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import  com.mitchell_international.VehicleApplication.model.Vehicle;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -19,19 +18,18 @@ import java.util.Optional;
 @Service
 public class VehicleService {
 
-
     @Autowired
     VehicleRepository vehicleRepository;
 
-
     public List<JsonNode> getVehicles() throws JsonProcessingException {
-
         Iterable<Vehicle> all = vehicleRepository.findAll();
+        if(all==null)
+            throw new VehiclesNotFoundException("No Vehicles found");
+
         Vehicle vehicleObject = new Vehicle();
         List<Vehicle> vehicleList = new ArrayList<>();
         ObjectMapper objectMapper = new ObjectMapper();
         List<JsonNode> jsonNodeList = new ArrayList<>();
-
         for(Vehicle vehicle : all){
             vehicleObject.setId(vehicle.getId());
             vehicleObject.setMake(vehicle.getMake());
@@ -41,16 +39,14 @@ public class VehicleService {
             JsonNode jsonNode = objectMapper.readTree(stocksListJsonString);
             jsonNodeList.add(jsonNode);
         }
-
-        if(jsonNodeList==null) {
-            throw new VehiclesNotFoundException("No Vehicles Found");
-        }
-
         return jsonNodeList;
     }
 
     public JsonNode getVehicleById(int id) throws JsonProcessingException {
         Optional<Vehicle> vehicleName = vehicleRepository.findById(id);
+
+       if(!vehicleName.isPresent())
+           throw new  VehicleNotFoundException("vehicle with id"+id+"does not exist");
         Vehicle vehicleObject = new Vehicle();
         vehicleObject.setId(vehicleName.get().getId());
         vehicleObject.setModel(vehicleName.get().getModel());
@@ -62,9 +58,7 @@ public class VehicleService {
         JsonNode vehicleJsonObject = objectMapper.readTree(vehicleObjectJsonString);
 
 
-        if(vehicleJsonObject==null){
-            throw new VehicleNotFoundException("This Vehicle is not found");
-        }
+
         return vehicleJsonObject;
 
 
@@ -72,10 +66,13 @@ public class VehicleService {
 
     public void createVehicles(Vehicle vehicle) {
         Vehicle vehicleObject = new Vehicle();
+
+
         vehicleObject.setId(vehicle.getId());
         vehicleObject.setMake(vehicle.getMake());
         vehicleObject.setModel(vehicle.getModel());
         vehicleObject.setYear(vehicle.getYear());
+
         vehicleRepository.save(vehicleObject);
     }
 
@@ -83,14 +80,21 @@ public class VehicleService {
 
     public void updateVehicles(Vehicle vehicle) {
         Optional<Vehicle> vehicleById = vehicleRepository.findById(vehicle.getId());
+        if(!vehicleById.isPresent())
+            throw new VehicleNotFoundException("Vehicle Not found");
+
           vehicleById.get().setMake(vehicle.getMake());
           vehicleById.get().setYear(vehicle.getYear());
           vehicleById.get().setModel(vehicle.getModel());
-
           vehicleRepository.save(vehicleById.get());
     }
 
     public void deleteVehiclesById(int id) {
-        vehicleRepository.deleteById(id);
+        Optional<Vehicle> vehicleById = vehicleRepository.findById(id);
+        if(vehicleById.isPresent())
+            vehicleRepository.deleteById(id);
+        else
+            throw new VehicleNotFoundException("Vehicle with this id "+id+"not found");
+
     }
 }
